reorganize to use babel-compile
NOTE: our babel preset does not support spread syntax.
