+++ .gitignore
@@ -27,3 +27,5 @@ coverage
 
 # Compiled binary addons (http://nodejs.org/api/addons.html)
 build/Release
+.test
+lib/
