+++ package.json
@@ -3,7 +3,10 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "test": "./test/runtests.sh"
+    "test": "mocha .test/*_test.js",
+    "compile": "babel-compile -p taskcluster src:lib test:.test",
+    "pretest": "npm run compile",
+    "install": "npm run compile"
   },
   "repository": {
     "type": "git",
@@ -12,7 +15,8 @@
   "dependencies": {
     "assume": "^1.2.4",
     "azure-entities": "^0.9.1",
-    "babel": "^5.0.9",
+    "babel-compile": "^2.0.0",
+    "babel-preset-taskcluster": "^3.0.0",
     "debug": "^2.1.3",
     "hawk": "^2.3.0",
     "lodash": "^3.6.0",
@@ -20,6 +24,7 @@
     "nock": "^4.0.0",
     "promise": "^6.1.0",
     "slugid": "^1.1.0",
+    "source-map-support": "^0.4.14",
     "taskcluster-client": "^1.6.3",
     "taskcluster-lib-api": "^3.1.0",
     "taskcluster-lib-app": "^0.8.10",
