+++ test/mocha.opts
@@ -1,4 +1,4 @@
 --ui tdd
---timeout 240s
+--timeout 30s
 --reporter spec
---compilers js:babel/register
+--require source-map-support/register
