--- test/runtests.sh
@@ -1,4 +1,0 @@
-#!/bin/bash -ve
-# USAGE: Run this file using `npm test` (must run from repository root)
-
-mocha test/*_test.js;
